Transactions - edit by user

+++ Wallet/Wallet.Data/Models/Transactions.cs
@@ -1,21 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
-
-// Code scaffolded by EF Core assumes nullable reference types (NRTs) are not used or disabled.
-// If you have enabled NRTs for your project, then un-comment the following line:
-// #nullable disable
 
 namespace Wallet.Data.Models
 {
     public partial class Transactions
     {
         public int Id { get; set; }
-        public float Amount { get; set; }
+        public double Amount { get; set; }
         public string Concept { get; set; }
         public DateTime Date { get; set; }
         public string Type { get; set; }
         public int AccountId { get; set; }
-
         public virtual Accounts Account { get; set; }
     }
 }
